Cleanup of preprocessors.
